Refactor: Move employee form to separate component
Moved the employee form from index.vue to its own dedicated component for better organization and maintainability
